SC-279: refactor ITextFrame
