+++ ShapeCrawler/Exceptions/PlaceholderCannotBeChangedException.cs
@@ -3,7 +3,7 @@
     internal class PlaceholderCannotBeChangedException : ShapeCrawlerException
     {
         internal PlaceholderCannotBeChangedException()
-            : base("The shape is a placeholder and cannot be changed on the slide level")
+            : base("The shape is a placeholder and cannot be changed on the Slide level")
         {
         }
     }
